rename to tsx files
