+++ src/index.js
@@ -46,7 +46,7 @@ import LandingEmailAd from './components/Ads/ShowcaseAds/LandingEmailAd';
 import LeadMarqueeCard from './components/Cards/LeadMarqueeCard';
 import LoadingCard from './components/Cards/LoadingCard';
 import LoadingCarousel from './components/Carousels/LoadingCarousel';
-import LoadingRelatedDocumentCard from './components/Cards/LoadingRelatedDocumentCard';
+import LoadingRelatedDocumentCard from './components/Cards/LoadingRelatedDocumentCard/LoadingRelatedDocumentCard';
 import MadeForYouCard from './components/Cards/MadeForYouCard';
 import MadeForYouCarousel from './components/Carousels/MadeForYouCarousel';
 import MarqueeCard from './components/Cards/MarqueeCard';
